requirements.txt
tensorflow-cpu
numpy
pillow
streamlit
streamlit-drawable-canvas